Working TV Boy support, refactor a2600 driver
