+++ src/mame/mess.flt
@@ -719,6 +719,7 @@ tv910.cpp
 tv912.cpp
 tv950.cpp
 tv990.cpp
+tvboy.cpp
 tvc.cpp
 tvgame.cpp
 tx0.cpp
